:recycle: Move modules to modules folder
